Allows starting individual branches expanded or collapsed
This update allows you to preset the expanded or collapsed class on
individual branches to give finer grain control over individual branch
status on start. This allows overriding the default startCollapsed
option on a per-branch level as needed.

I needed this so that I could ‘remember’ the state of individual
branches and restore them on page load.